# docs/openapi/generated/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NIPP API",
    "version": "0.1.0",
    "description": "Generated by `npm run docs:generate` from JSDoc headers in app/api/ (see scripts/generate-openapi.mjs). Request/response schemas are not yet modelled — they will appear as zod-to-openapi definitions are added."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "BetterAuth session cookie. Set the cookie in Swagger UI using Authorize."
      }
    }
  },
  "paths": {
...
    "/api/admin/organizations/{orgId}": {
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "archive; PENDING orgs are hard-deleted",
        "description": "DELETE — archive; PENDING orgs are hard-deleted\n\nSuper Admin only. Writes are rate-limited per session. Runs under a verified\ntarget-org context (RLS Phase 3); every mutation is audit-logged and\ninvalidates the `org` cache tag. DELETE state machine: PENDING → deleted,\nACTIVE/SUSPENDED → ARCHIVED (terminal), ARCHIVED → 400.\n\nRequires super admin session.\n\nHandler: `app/api/admin/organizations/[orgId]/route.ts`.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Unspecified — response schema not yet documented (add a zod schema + JSDoc header to fill this in)."
          }
        }
      },
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "organization details by ID (cached via tag `org`)",
        "description": "GET — organization details by ID (cached via tag `org`)\n\nSuper Admin only. Writes are rate-limited per session. Runs under a verified\ntarget-org context (RLS Phase 3); every mutation is audit-logged and\ninvalidates the `org` cache tag. DELETE state machine: PENDING → deleted,\nACTIVE/SUSPENDED → ARCHIVED (terminal), ARCHIVED → 400.\n\nRequires super admin session.\n\nHandler: `app/api/admin/organizations/[orgId]/route.ts`.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Unspecified — response schema not yet documented (add a zod schema + JSDoc header to fill this in)."
          }
        }
      },
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "update name/slug/status (slug uniqueness enforced)",
        "description": "PATCH — update name/slug/status (slug uniqueness enforced)\n\nSuper Admin only. Writes are rate-limited per session. Runs under a verified\ntarget-org context (RLS Phase 3); every mutation is audit-logged and\ninvalidates the `org` cache tag. DELETE state machine: PENDING → deleted,\nACTIVE/SUSPENDED → ARCHIVED (terminal), ARCHIVED → 400.\n\nRequires super admin session.\n\nHandler: `app/api/admin/organizations/[orgId]/route.ts`.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Unspecified — response schema not yet documented (add a zod schema + JSDoc header to fill this in)."
          }
        }
      }
    },
...
  }
}